// src/app/_restProducto/model/productosInner.ts
/**
 * Servicio Productos Kallsonys
 * Servicio para la gestión de productos en  OMS Kallsonys
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Cambios por cada uno de los servicios
 */
export interface ProductosInner { 
    nombre?: string;
    /**
     * información de devuelta.
     */
    idProducto?: number;
    valorBase?: number;
    descripcion?: string;
    idcategoria?: number;
    rutaImagen?: string;
    estado?: string;
}