// src/app/_restProducto/model/productoRsType.ts
/**
 * Servicio Productos Kallsonys
 * Servicio para la gestión de productos en  OMS Kallsonys
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Productos } from './productos';
import { StatusType } from './statusType';


/**
 * Respuesta genérica.
 */
export interface ProductoRsType { 
    status?: StatusType;
    productos?: Productos;
}